backend/requirements.txt
debugpy
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt
gunicorn>=21.2.0
psycopg2-binary
celery
django-extensions